extract stderr splitting to separate pkg so runner can use @packages/error
