+++ packages/data-context/src/DataContext.ts
@@ -38,7 +38,8 @@ import type { IncomingHttpHeaders } from 'http'
 import type { App as ElectronApp } from 'electron'
 import { globalPubSub } from '.'
 import { ProjectLifecycleManager } from './data/ProjectLifecycleManager'
-import { CypressError, logError } from '@packages/errors'
+import { logError } from '@packages/stderr-filtering'
+import type { CypressError } from '@packages/errors'
 import { resetIssuedWarnings } from '@packages/config'
 
 const IS_DEV_ENV = process.env.CYPRESS_INTERNAL_ENV !== 'production'
