+++ packages/errors/src/index.ts
@@ -11,11 +11,3 @@ export * from './errors'
 export * from './errorTypes'
 
 export default errorsApi
-
-export { logError, START_TAG, END_TAG } from './log'
-
-export { FilterTaggedContent } from './stderrSplitting/FilterTaggedContent'
-
-export { FilterPrefixedContent } from './stderrSplitting/FilterPrefixedContent'
-
-export { WriteToDebug } from './stderrSplitting/WriteToDebug'
