+++ packages/errors/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "extends": "../ts/tsconfig.json",
   "include": [
-    "src",
+    "src", "../stderr-filtering/lib/logError.ts", "../stderr-filtering/lib/stderrSplitting",
   ],
   "compilerOptions": {
     "allowJs": false,
