+++ packages/graphql/package.json
@@ -20,6 +20,7 @@
   "dependencies": {
     "@graphql-tools/delegate": "8.2.1",
     "@graphql-tools/wrap": "8.1.1",
+    "@packages/stderr-filtering": "0.0.0-development",
     "@urql/core": "2.4.4",
     "chalk": "4.1.2",
     "cors": "2.8.5",
