+++ packages/server/package.json
@@ -34,6 +34,7 @@
     "@cypress/webpack-preprocessor": "0.0.0-development",
     "@ffmpeg-installer/ffmpeg": "1.1.0",
     "@packages/icons": "0.0.0-development",
+    "@packages/stderr-filtering": "0.0.0-development",
     "@packages/telemetry": "0.0.0-development",
     "@types/mime": "^3.0.1",
     "ansi_up": "5.0.0",
