+++ packages/stderr-filtering/lib/FilterPrefixedContent.ts
