+++ packages/stderr-filtering/lib/FilterTaggedContent.ts
