+++ packages/stderr-filtering/lib/LineDecoder.ts
