+++ packages/stderr-filtering/lib/WriteToDebug.ts
