+++ packages/stderr-filtering/lib/index.ts
@@ -1,0 +1,7 @@
+export { FilterTaggedContent } from './FilterTaggedContent'
+
+export { FilterPrefixedContent } from './FilterPrefixedContent'
+
+export { WriteToDebug } from './WriteToDebug'
+
+export { logError, START_TAG, END_TAG } from './logError'
