+++ packages/stderr-filtering/lib/logError.ts
