+++ packages/stderr-filtering/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "@packages/stderr-filtering",
+  "version": "0.0.0-development",
+  "private": true,
+  "main": "dist/index.js",
+  "scripts": {
+    "build": "tsc",
+    "watch": "tsc --watch"
+  },
+  "types": "dist/index.d.ts"
+}
